fix: resolve remaining review feedback
