+++ packages/riviere-cli/src/features/extract/domain/detect-connections-per-module.spec.ts
@@ -89,7 +89,7 @@ describe('ExtractionProject.extractDraftComponents', () => {
     )
   })
 
-  it('returns the shared-core draft result unchanged', () => {
+  it('returns the shared-core draft result object unchanged', () => {
     mockExtractInto.mockReturnValueOnce({
       kind: 'draftOnly',
       components: [
